requirements.txt
asgiref==3.7.2
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.0.0
